requirements.txt
hdwallet>=2.2.1
cryptography>=41.0.0,<42.0.0
fastapi==0.79.0
pydantic==1.9.2
python-dotenv==0.19.0
rich>=12.0.0,<13.0.0
click>=8.0.0,<9.0.0
colorama>=0.4.0,<0.5.0
python-bitcoinrpc>=1.0
uvicorn>=0.15.0
flask>=2.0.0
sqlalchemy>=1.4.0
aiofiles>=0.8.0
python-multipart>=0.0.5 